Refactor database connection URL and update Dockerfile
This commit updates the database connection URL in the app.js file to use a local MongoDB instance instead of a remote one. Additionally, it adds a new start script in the package.json file to run the application with the "node" command. The Dockerfile has also been added to the repository, which sets up the application in a Docker container.